Added AI (layout scripts)
Added 1 script used for running away (for testing).
then I added 4 new scripts. With those scripts we can use state machines. For now there are only two states: Normal & RunningAway.
I only made running away. When the distance between the animal and player is less than a float, then it'll run away. When the distance is greater then the float, then the animal will stop. So it won't continue running, it'll go back to the Normal State.

These states won't work in the game, because i deleted the changes to the scene (since it would cause problems with mergin)