Added perches and some other changes
There is now a timer with a random number that will make the bird fly to a perch.
Also, I also added something that can be used to have more birds use these 'areas'. I made it so that the tree perches can't be used when it's occupied.

(I didn't test everything a lot, so there can be some issues here and there...)